Move some methods to block module
